+++ src/Backend/SgateShopgatePlugin/Components/Cart.php
@@ -366,6 +366,25 @@ class Cart
         if ($voucher) {
             $response['promotionVouchers'] = json_encode($this->session->get('promotionVouchers'));
 
+            // Check if we have an error while adding the voucher and that error has a message we can parse
+            if ($response['addVoucher'] && @$response['addVoucher']['sErrorMessages'][0]) {
+                // If the error message has a number in it, it's likely that the minimum charge wasn't reached
+                if (preg_match('/\b\d+[\,\.]?\d*\b/', $response['addVoucher']['sErrorMessages'][0])) {
+
+                    // Get that voucher from the db
+                    $voucherDetails = Shopware()->Db()->fetchRow(
+                        'SELECT minimumcharge FROM s_emarketing_vouchers WHERE vouchercode = ? LIMIT 1',
+                        strtolower(trim(stripslashes($code)))
+                    );
+
+                    // And add the minimum charge to the response if it has one
+                    if ($voucherDetails && $voucherDetails['minimumcharge'] > 0) {
+                        $response['addVoucher']['minimumcharge'] = $voucherDetails['minimumcharge'];
+                    }
+                }
+            }
+
+
             $httpResponse->setBody(html_entity_decode(json_encode($response)));
             $httpResponse->sendResponse();
             exit();